feat(build): reorganized lib into its own custom-element folder - just to cleanup the library, move the "aurelia-slickgrid" and "slick-pagination" into a new "custom-elements" folder and in that folder create the test folder as well.
