+++ src/aurelia-slickgrid/custom-elements/aurelia-slickgrid.html
