+++ src/aurelia-slickgrid/custom-elements/slick-pagination.html
