+++ src/aurelia-slickgrid/services/utilities.ts
@@ -127,7 +127,7 @@ export function decimalFormatted(input: number | string, minDecimal?: number, ma
 
   // do we want to display our number with a custom separator in each thousand position
   if (thousandSeparator) {
-    amount = thousandSeparatorFormatted(amount, thousandSeparator);
+    amount = thousandSeparatorFormatted(amount, thousandSeparator) || '';
   }
 
   // when using a separator that is not a dot, replace it with the new separator
@@ -148,20 +148,20 @@ export function formatNumber(input: number | string, minDecimal?: number, maxDec
   if (calculatedValue < 0) {
     const absValue = Math.abs(calculatedValue);
     if (displayNegativeNumberWithParentheses) {
-      if (!isNaN(minDecimal) || !isNaN(maxDecimal)) {
+      if (!isNaN(minDecimal as number) || !isNaN(maxDecimal as number)) {
         return `(${symbolPrefix}${decimalFormatted(absValue, minDecimal, maxDecimal, decimalSeparator, thousandSeparator)}${symbolSuffix})`;
       }
       const formattedValue = thousandSeparatorFormatted(`${absValue}`, thousandSeparator);
       return `(${symbolPrefix}${formattedValue}${symbolSuffix})`;
     } else {
-      if (!isNaN(minDecimal) || !isNaN(maxDecimal)) {
+      if (!isNaN(minDecimal as number) || !isNaN(maxDecimal as number)) {
         return `-${symbolPrefix}${decimalFormatted(absValue, minDecimal, maxDecimal, decimalSeparator, thousandSeparator)}${symbolSuffix}`;
       }
       const formattedValue = thousandSeparatorFormatted(`${absValue}`, thousandSeparator);
       return `-${symbolPrefix}${formattedValue}${symbolSuffix}`;
     }
   } else {
-    if (!isNaN(minDecimal) || !isNaN(maxDecimal)) {
+    if (!isNaN(minDecimal as number) || !isNaN(maxDecimal as number)) {
       return `${symbolPrefix}${decimalFormatted(input, minDecimal, maxDecimal, decimalSeparator, thousandSeparator)}${symbolSuffix}`;
     }
     const formattedValue = thousandSeparatorFormatted(`${input}`, thousandSeparator);
